+++ src/app/features/search/models/search-request.model.ts
@@ -1,4 +1,4 @@
-import { ActivityTag } from '../../../core/models/activity-tag.model';
+import { ActivityTag } from '../../activity-tag/models/activity-tag.model';
 import { Destination } from './destination.model';
 
 export interface SearchRequest {
